metadata: Move Validate() and Populate() logic to pkg/metadata/v1
