+++ pkg/oci/build_metadata.go
@@ -24,7 +24,6 @@ import (
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 
-	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/run/types"
 	metadatav1 "github.com/inspektor-gadget/inspektor-gadget/pkg/metadata/v1"
 )
 
@@ -79,7 +78,7 @@ func validateMetadataFile(ctx context.Context, opts *BuildGadgetImageOpts) error
 		return fmt.Errorf("loading spec: %w", err)
 	}
 
-	return types.Validate(metadata, spec)
+	return metadata.Validate(spec)
 }
 
 func createOrUpdateMetadataFile(ctx context.Context, opts *BuildGadgetImageOpts) error {
@@ -108,14 +107,14 @@ func createOrUpdateMetadataFile(ctx context.Context, opts *BuildGadgetImageOpts)
 		log.Debugf("Metadata file found, updating it")
 
 		// TODO: this validation could be softer, just printing warnings
-		if err := types.Validate(metadata, spec); err != nil {
+		if err := metadata.Validate(spec); err != nil {
 			return fmt.Errorf("metadata file is wrong, fix it before continuing: %w", err)
 		}
 	} else {
 		log.Debug("Metadata file not found, generating it")
 	}
 
-	if err := types.Populate(metadata, spec); err != nil {
+	if err := metadata.Populate(spec); err != nil {
 		return fmt.Errorf("handling trace maps: %w", err)
 	}
